+++ spec/unit/hanami/assets/helpers_spec.rb
@@ -2,6 +2,11 @@ describe Hanami::Assets::Helpers do
   let(:view)    { ImageHelperView.new({}, {}) }
   let(:cdn_url) { 'https://bookshelf.cdn-example.com' }
 
+  before do
+    view.class.assets_configuration.load!
+    Thread.current[:__hanami_assets] = nil
+  end
+
   after do
     view.class.assets_configuration.reset!
   end
@@ -75,6 +80,45 @@ describe Hanami::Assets::Helpers do
         expect(actual).to eq(%(<script src="#{cdn_url}/assets/feature-a.js" type="text/javascript"></script>))
       end
     end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "includes asset in push promise assets" do
+        DefaultView.new.javascript("feature-a")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets.fetch("/assets/feature-a.js")).to eq(as: :script, crossorigin: false)
+      end
+
+      it "includes crossorigin asset in push promise assets" do
+        DefaultView.new.javascript("https://assets.hanamirb.org/assets/framework.js")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets.fetch("https://assets.hanamirb.org/assets/framework.js")).to eq(as: :script, crossorigin: true)
+      end
+
+      it "allows asset exclusion from push promise assets" do
+        actual = DefaultView.new.javascript("feature-b", push: false)
+        expect(actual).to eq(%(<script src="/assets/feature-b.js" type="text/javascript"></script>))
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "includes multiple assets in push promise assets" do
+        DefaultView.new.javascript("feature-c", "feature-d")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets.fetch("/assets/feature-c.js")).to eq(as: :script, crossorigin: false)
+        expect(assets.fetch("/assets/feature-d.js")).to eq(as: :script, crossorigin: false)
+      end
+
+      it "allows the exclusion of multiple assets from push promise assets" do
+        DefaultView.new.javascript("feature-e", "feature-f", push: false)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+    end
   end
 
   describe '#stylesheet' do
@@ -129,6 +173,38 @@ describe Hanami::Assets::Helpers do
         expect(actual).to eq(%(<link href="#{cdn_url}/assets/main.css" type="text/css" rel="stylesheet">))
       end
     end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "includes asset in push promise assets" do
+        DefaultView.new.stylesheet("main")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets.fetch("/assets/main.css")).to eq(as: :style, crossorigin: false)
+      end
+
+      it "allows asset exclusion from push promise assets" do
+        actual = DefaultView.new.stylesheet("fonts", push: false)
+        expect(actual).to eq(%(<link href="/assets/fonts.css" type="text/css" rel="stylesheet">))
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "includes multiple assets in push promise assets" do
+        DefaultView.new.stylesheet("framework", "styles")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets.fetch("/assets/framework.css")).to eq(as: :style, crossorigin: false)
+        expect(assets.fetch("/assets/styles.css")).to eq(as: :style, crossorigin: false)
+      end
+
+      it "allows the exclusion of multiple assets from push promise assets" do
+        DefaultView.new.stylesheet("framework", "styles", push: false)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+    end
   end
 
   describe 'image' do
@@ -167,6 +243,31 @@ describe Hanami::Assets::Helpers do
         expect(actual).to eq(%(<img src="#{cdn_url}/assets/application.jpg" alt="Application">))
       end
     end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't include asset in push promise assets" do
+        view.image("application.jpg")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "allows assets inclusion in push promise assets" do
+        actual = view.image("application.jpg", push: true).to_s
+        expect(actual).to eq(%(<img src="/assets/application.jpg" alt="Application">))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/application.jpg")).to eq(as: :image, crossorigin: false)
+      end
+
+      it "allows crossorigin assets inclusion in push promise assets" do
+        actual = view.image("https://assets.hanamirb.org/assets/application.jpg", push: true).to_s
+        expect(actual).to eq(%(<img src="https://assets.hanamirb.org/assets/application.jpg" alt="Application">))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("https://assets.hanamirb.org/assets/application.jpg")).to eq(as: :image, crossorigin: true)
+      end
+    end
   end
 
   describe '#favicon' do
@@ -198,6 +299,31 @@ describe Hanami::Assets::Helpers do
       it 'returns absolute url for href attribute' do
         actual = view.favicon.to_s
         expect(actual).to eq(%(<link href="#{cdn_url}/assets/favicon.ico" rel="shortcut icon" type="image/x-icon">))
+      end
+    end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't include asset in push promise assets" do
+        view.favicon
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "allows assets inclusion in push promise assets" do
+        actual = view.favicon("favicon.ico", push: true).to_s
+        expect(actual).to eq(%(<link href="/assets/favicon.ico" rel="shortcut icon" type="image/x-icon">))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/favicon.ico")).to eq(as: :image, crossorigin: false)
+      end
+
+      it "allows crossorigin assets inclusion in push promise assets" do
+        actual = view.favicon("https://assets.hanamirb.org/assets/favicon.ico", push: true).to_s
+        expect(actual).to eq(%(<link href="https://assets.hanamirb.org/assets/favicon.ico" rel="shortcut icon" type="image/x-icon">))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("https://assets.hanamirb.org/assets/favicon.ico")).to eq(as: :image, crossorigin: true)
       end
     end
   end
@@ -268,6 +394,55 @@ describe Hanami::Assets::Helpers do
         expect(actual).to eq(%(<video src="#{cdn_url}/assets/movie.mp4"></video>))
       end
     end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't include asset in push promise assets" do
+        view.video("movie.mp4")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "allows asset inclusion in push promise assets" do
+        actual = view.video("movie.mp4", push: true).to_s
+        expect(actual).to eq(%(<video src="/assets/movie.mp4"></video>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/movie.mp4")).to eq(as: :video, crossorigin: false)
+      end
+
+      it "allows crossorigin asset inclusion in push promise assets" do
+        actual = view.video("https://assets.hanamirb.org/assets/movie.mp4", push: true).to_s
+        expect(actual).to eq(%(<video src="https://assets.hanamirb.org/assets/movie.mp4"></video>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("https://assets.hanamirb.org/assets/movie.mp4")).to eq(as: :video, crossorigin: true)
+      end
+
+      it "allows asset inclusion in push promise assets when using block syntax" do
+        actual = view.video("movie.mp4", push: true) do
+          "Your browser does not support the video tag"
+        end.to_s
+
+        expect(actual).to eq(%(<video src="/assets/movie.mp4">\nYour browser does not support the video tag\n</video>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/movie.mp4")).to eq(as: :video, crossorigin: false)
+      end
+
+      it "allows asset inclusion in push promise assets when using block syntax and source tags" do
+        actual = view.video do
+          text "Your browser does not support the video tag"
+          source src: view.asset_path("movie.mp4", push: :video), type: "video/mp4"
+          source src: view.asset_path("movie.ogg"), type: "video/ogg"
+        end.to_s
+
+        expect(actual).to eq(%(<video>\nYour browser does not support the video tag\n<source src="/assets/movie.mp4" type="video/mp4">\n<source src="/assets/movie.ogg" type="video/ogg">\n</video>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/movie.mp4")).to eq(as: :video, crossorigin: false)
+      end
+    end
   end
 
   describe '#audio' do
@@ -336,6 +511,55 @@ describe Hanami::Assets::Helpers do
         expect(actual).to eq(%(<audio src="#{cdn_url}/assets/song.ogg"></audio>))
       end
     end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't include asset in push promise assets" do
+        view.audio("song.ogg")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "allows asset inclusion in push promise assets" do
+        actual = view.audio("song.ogg", push: true).to_s
+        expect(actual).to eq(%(<audio src="/assets/song.ogg"></audio>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/song.ogg")).to eq(as: :audio, crossorigin: false)
+      end
+
+      it "allows crossorigin asset inclusion in push promise assets" do
+        actual = view.audio("https://assets.hanamirbg.org/assets/song.ogg", push: true).to_s
+        expect(actual).to eq(%(<audio src="https://assets.hanamirbg.org/assets/song.ogg"></audio>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("https://assets.hanamirbg.org/assets/song.ogg")).to eq(as: :audio, crossorigin: true)
+      end
+
+      it "allows asset inclusion in push promise assets when using block syntax" do
+        actual = view.audio("song.ogg", push: true) do
+          "Your browser does not support the audio tag"
+        end.to_s
+
+        expect(actual).to eq(%(<audio src="/assets/song.ogg">\nYour browser does not support the audio tag\n</audio>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/song.ogg")).to eq(as: :audio, crossorigin: false)
+      end
+
+      it "allows asset inclusion in push promise assets when using block syntax and source tags" do
+        actual = view.audio do
+          text "Your browser does not support the audio tag"
+          source src: view.asset_path("song.ogg", push: :audio), type: "audio/ogg"
+          source src: view.asset_path("song.wav"), type: "audio/wav"
+        end.to_s
+
+        expect(actual).to eq(%(<audio>\nYour browser does not support the audio tag\n<source src="/assets/song.ogg" type="audio/ogg">\n<source src="/assets/song.wav" type="audio/wav">\n</audio>))
+
+        assets = Thread.current[:__hanami_assets]
+        expect(assets.fetch("/assets/song.ogg")).to eq(as: :audio, crossorigin: false)
+      end
+    end
   end
 
   describe '#asset_path' do
@@ -349,11 +573,6 @@ describe Hanami::Assets::Helpers do
       expect(result).to eq('http://assets.hanamirb.org/assets/application.css')
     end
 
-    it 'adds source to HTTP/2 PUSH PROMISE list' do
-      view.asset_path('dashboard.js')
-      expect(Thread.current[:__hanami_assets]).to include('/assets/dashboard.js')
-    end
-
     describe 'cdn mode' do
       before do
         activate_cdn_mode!
@@ -362,6 +581,39 @@ describe Hanami::Assets::Helpers do
       it 'returns absolute url' do
         result = view.asset_path('application.js')
         expect(result).to eq('https://bookshelf.cdn-example.com/assets/application.js')
+      end
+    end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't add into assets list by default" do
+        view.asset_path("dashboard.js")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "adds asset into assets list" do
+        view.asset_path("dashboard.js", push: true)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch('/assets/dashboard.js')).to eq(as: nil, crossorigin: false)
+      end
+
+      it "allows to specify asset type" do
+        view.asset_path("video.mp4", push: :video)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch('/assets/video.mp4')).to eq(as: :video, crossorigin: false)
+      end
+
+      it "allows to link crossorigin asset" do
+        view.asset_path("https://assets.hanamirb.org/assets/video.mp4", push: :video)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch("https://assets.hanamirb.org/assets/video.mp4")).to eq(as: :video, crossorigin: true)
       end
     end
   end
@@ -381,11 +633,6 @@ describe Hanami::Assets::Helpers do
       expect(result).to eq('http://assets.hanamirb.org/assets/application.css')
     end
 
-    it 'adds source to HTTP/2 PUSH PROMISE list' do
-      view.asset_url('metrics.js')
-      expect(Thread.current[:__hanami_assets]).to include('http://localhost:2300/assets/metrics.js')
-    end
-
     describe 'cdn mode' do
       before do
         activate_cdn_mode!
@@ -394,6 +641,39 @@ describe Hanami::Assets::Helpers do
       it 'still returns absolute url' do
         result = view.asset_url('application.js')
         expect(result).to eq('https://bookshelf.cdn-example.com/assets/application.js')
+      end
+    end
+
+    context "HTTP/2 PUSH PROMISE" do
+      it "doesn't add into assets list by default" do
+        view.asset_url("metrics.js")
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be(nil)
+      end
+
+      it "adds asset into assets list" do
+        view.asset_url("metrics.js", push: true)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch("http://localhost:2300/assets/metrics.js")).to eq(as: nil, crossorigin: false)
+      end
+
+      it "allows to specify asset type" do
+        view.asset_url("metrics.js", push: :script)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch("http://localhost:2300/assets/metrics.js")).to eq(as: :script, crossorigin: false)
+      end
+
+      it "allows to link crossorigin asset" do
+        view.asset_url("https://assets.hanamirb.org/assets/metrics.js", push: :script)
+        assets = Thread.current[:__hanami_assets]
+
+        expect(assets).to be_kind_of(Hash)
+        expect(assets.fetch("https://assets.hanamirb.org/assets/metrics.js")).to eq(as: :script, crossorigin: true)
       end
     end
   end