+++ spec/unit/hanami/assets/version_spec.rb
@@ -1,5 +1,5 @@
 describe "Hanami::Assets::VERSION" do
   it "exposes version" do
-    expect(Hanami::Assets::VERSION).to eq("1.1.1")
+    expect(Hanami::Assets::VERSION).to eq("1.2.0")
   end
 end