+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Hanami::Assets
 Assets management for Ruby web applications
 
+## v1.2.0 - 2018-04-11
+
+## v1.2.0.rc2 - 2018-04-06
+
+## v1.2.0.rc1 - 2018-03-30
+
+## v1.2.0.beta2 - 2018-03-23
+
+## v1.2.0.beta1 - 2018-02-28
+### Added
+- [Luca Guidi] Collect assets informations for Early Hints (103)
+- [Luca Guidi] Send automatically javascripts and stylesheets via Push Promise / Early Hints
+- [Luca Guidi] Add the ability to send audio, video, and generic assets for Push Promise / Early Hints
+
 ## v1.1.1 - 2018-02-27
 ### Added
 - [Luca Guidi] Official support for Ruby: MRI 2.5